remove DecodedViewerFinalSettings
