+++ cmd/frontend/graphqlbackend/search.go
@@ -7,10 +7,10 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
 	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/search"
 	"github.com/sourcegraph/sourcegraph/internal/search/client"
 	"github.com/sourcegraph/sourcegraph/internal/search/job/jobutil"
+	"github.com/sourcegraph/sourcegraph/internal/settings"
 	"github.com/sourcegraph/sourcegraph/internal/trace"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 	"github.com/sourcegraph/sourcegraph/schema"
@@ -30,7 +30,7 @@ type SearchImplementer interface {
 
 // NewBatchSearchImplementer returns a SearchImplementer that provides search results and suggestions.
 func NewBatchSearchImplementer(ctx context.Context, logger log.Logger, db database.DB, enterpriseJobs jobutil.EnterpriseJobs, args *SearchArgs) (_ SearchImplementer, err error) {
-	settings, err := DecodedViewerFinalSettings(ctx, db)
+	settings, err := settings.CurrentUserFinal(ctx, db)
 	if err != nil {
 		return nil, err
 	}
@@ -76,8 +76,6 @@ type searchResolver struct {
 	enterpriseJobs jobutil.EnterpriseJobs
 }
 
-var MockDecodedViewerFinalSettings *schema.Settings
-
 // DecodedViewerFinalSettings returns the final (merged) settings for the viewer
 func DecodedViewerFinalSettings(ctx context.Context, db database.DB) (_ *schema.Settings, err error) {
 	tr, ctx := trace.New(ctx, "decodedViewerFinalSettings", "")
@@ -85,14 +83,5 @@ func DecodedViewerFinalSettings(ctx context.Context, db database.DB) (_ *schema.
 		tr.SetError(err)
 		tr.Finish()
 	}()
-	if MockDecodedViewerFinalSettings != nil {
-		return MockDecodedViewerFinalSettings, nil
-	}
-
-	cascade, err := newSchemaResolver(db, gitserver.NewClient(), nil).ViewerSettings(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return cascade.finalTyped(ctx)
+	return nil, nil
 }
