+++ enterprise/cmd/frontend/internal/compute/streaming/compute.go
@@ -7,7 +7,6 @@ import (
 	"github.com/sourcegraph/log"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/compute"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/search"
@@ -15,6 +14,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/search/job/jobutil"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 	"github.com/sourcegraph/sourcegraph/internal/search/streaming"
+	"github.com/sourcegraph/sourcegraph/internal/settings"
 )
 
 func toComputeResult(ctx context.Context, cmd compute.Command, match result.Match) (out []compute.Result, _ error) {
@@ -67,7 +67,7 @@ func NewComputeStream(ctx context.Context, logger log.Logger, db database.DB, en
 		}
 	})
 
-	settings, err := graphqlbackend.DecodedViewerFinalSettings(ctx, db)
+	settings, err := settings.CurrentUserFinal(ctx, db)
 	if err != nil {
 		close(eventsC)
 		close(errorC)
